# docs/swagger.yaml
basePath: /
definitions:
  products.Product:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      update_at:
        type: string
    type: object
  users.Users:
    properties:
      address:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      password:
        type: string
      permission_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gin Swagger Example API
  version: "1.0"
paths:
  /product/all:
    get:
      consumes:
      - application/json
      description: Get all products from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/products.Product'
            type: array
      summary: Get all products
      tags:
      - product
  /product/all/filter:
    get:
      consumes:
      - application/json
      description: Get a list of products with pagination
      parameters:
      - description: Product name
        in: query
        name: pname
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/products.Product'
            type: array
      summary: Get products
      tags:
      - product
  /user/all:
    get:
      consumes:
      - application/json
      description: Get all User from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.Users'
            type: array
      summary: Get all User
      tags:
      - User
swagger: "2.0"